Rework items y creacion player

+++ Assets/Scripts/Events/EventData.cs
@@ -2,12 +2,12 @@ namespace Game.Events
 {
     public class EventData
     {
-        public int amount;
+        public float amount;
         public dynamic source;
         public dynamic destination;
         public dynamic type;
 
-        public EventData(int amount = 0, dynamic source = null, dynamic destination = null, dynamic type = null)
+        public EventData(float amount = 0, dynamic source = null, dynamic destination = null, dynamic type = null)
         {
             this.amount = amount;
             this.source = source;
